+++ src/main/java/io/github/astrapi69/swing/tree/renderer/GenericTreeNodeCellRenderer.java
@@ -30,17 +30,17 @@ import org.apache.commons.lang3.StringUtils;
 
 import io.github.astrapi69.icon.ImageIconFactory;
 import io.github.astrapi69.icon.StringIcon;
-import io.github.astrapi69.swing.tree.JXTreeElement;
+import io.github.astrapi69.swing.tree.GenericTreeElement;
 import io.github.astrapi69.tree.TreeNode;
 
-public class JXTreeNodeCellRenderer extends TreeNodeCellRenderer<JXTreeElement>
+public class GenericTreeNodeCellRenderer<T> extends TreeNodeCellRenderer<GenericTreeElement<T>>
 {
 
-	protected JLabel initialize(TreeNode<JXTreeElement> userObject)
+	protected JLabel initialize(TreeNode<GenericTreeElement<T>> userObject)
 	{
-		TreeNode<JXTreeElement> treeNode = userObject;
+		TreeNode<GenericTreeElement<T>> treeNode = userObject;
 		String displayValue = treeNode.getDisplayValue();
-		JXTreeElement value = treeNode.getValue();
+		GenericTreeElement<T> value = treeNode.getValue();
 		if (value != null)
 		{
 			String iconPath = value.getIconPath();