+++ src/test/java/io/github/astrapi69/swing/tree/panel/model/TreeNodeMutableTreeNode.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.tree.panel;
+package io.github.astrapi69.swing.tree.panel.model;
 
 import java.io.Serializable;
 import java.util.Collections;